pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "s4_dx7"
version = "0.1.0"
description = "Project description"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
dbt-core = "^1.7.4"
dbt-duckdb = "^1.7.0"
ray = "^2.9.0"
pedalboard = "^0.8.7"
mido = "^1.3.2"
tqdm = "^4.66.1"
pydub = "^0.25.1"
pyarrow = "^14.0.2"
pandas = "^2.1.4"
hydra-core = "^1.3.2"
pytorch-lightning = "2.0.4"
transformers = "^4.36.2"
wandb = "^0.16.2"
einops = "^0.7.0"
rich = "^13.7.0"
torchaudio = "^2.1.2"
torchvision = "^0.16.2"
torchtext = "^0.16.2"
datasets = "^2.16.1"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.2"
torch = "^2.1.2"
natsort = "^8.4.0"
grpcio = "^1.60.0"
librosa = "^0.10.1"
bitstruct = "^8.19.0"

[tool.poetry.group.vis.dependencies]
plotly = "^5.18.0"
nbformat = "^5.9.2"
ipykernel = "^6.28.0"


[tool.poetry.group.debug.dependencies]
ray = {extras = ["default"], version = "^2.9.0"}


[tool.poetry.group.nb.dependencies]
librosa = "^0.10.1"

